+++ resources.tf
@@ -9,9 +9,9 @@ resource "aws_vpc" "development" {
 }
 
 resource "aws_subnet" "subnet1" {
-  cidr_block = cidrsubnet(aws_vpc.development.cidr_block, 1, 1)
-  vpc_id     = aws_vpc.development.id
-  provider   = aws.east
+  cidr_block        = cidrsubnet(aws_vpc.development.cidr_block, 3, 1)
+  vpc_id            = aws_vpc.development.id
+  availability_zone = "us-east-1a"
   tags = {
     Name = "devsubnet1"
   }
@@ -22,11 +22,14 @@ resource "aws_security_group" "subnetsecurity" {
 
   ingress {
     cidr_blocks = [
-      aws_vpc.development.cidr_block
+      "0.0.0.0/0"
     ]
 
-    from_port = 80
-    to_port   = 80
+    from_port = var.http_port
+    to_port   = var.http_port
     protocol  = "tcp"
+  }
+  tags = {
+    Name = "HTTP ingress"
   }
 }