+++ servers.tf
@@ -14,11 +14,19 @@ data "aws_ami" "ubuntu" {
 }
 
 resource "aws_instance" "devserver00" {
-  ami           = data.aws_ami.ubuntu.id
-  instance_type = "t2.micro"
+  ami                         = data.aws_ami.ubuntu.id
+  instance_type               = "t2.micro"
+  vpc_security_group_ids      = [aws_security_group.subnetsecurity.id]
+  associate_public_ip_address = true
+
+  user_data = <<-EOF
+              #!/bin/bash
+              echo "Hello, World" > index.html
+              nohup busybox httpd -f -p "${var.http_port}" &
+              EOF
 
   tags = {
     Name = "devserver00"
   }
-  subnet_id = "aws_subnet.subnet1.id"
+  subnet_id = aws_subnet.subnet1.id
 }