+++ variables.tf
@@ -1,0 +1,5 @@
+variable "http_port" {
+  default     = "8080"
+  type        = number
+  description = "Default HTTP port"
+}